chore(fixed_point): move cordic implementation to its header file
